Refactor username sanitization: move sanitize_username function to string_utils and remove username.py
